remove dom interaction logic into file
